changed CRTP to pure virtual methods
