+++ src/ros/rover/drive/drive_controllers/nova_drive_controller_base/CMakeLists.txt
@@ -41,10 +41,11 @@ generate_parameter_library(nova_drive_controller_base_parameters
 )
 
 add_library(${PROJECT_NAME} SHARED
+  src/nova_drive_controller_base.cpp
   src/odometry.cpp
 )
-target_include_directories(${PROJECT_NAME}
-  PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include/>
+target_include_directories(${PROJECT_NAME} PUBLIC
+  $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include/>
   $<INSTALL_INTERFACE:include>)
 ament_target_dependencies(${PROJECT_NAME} ${THIS_PACKAGE_INCLUDE_DEPENDS})
 target_link_libraries(${PROJECT_NAME}
